extract ClientConnection from Hocuspocus.ts
